requirements.txt
# Core API & UI
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
streamlit==1.31.0
requests==2.31.0

# Data & Visualization
pandas==2.2.0
numpy==1.26.3
plotly==5.18.0
matplotlib==3.8.2

# ML & NLP
torch==2.0.1
openai-whisper==20231117
sentence-transformers==2.2.2
spacy==3.7.2
language-tool-python==2.7.1
vaderSentiment==3.3.2
nltk==3.8.1
transformers==4.30.2
tokenizers==0.13.3
huggingface-hub==0.20.3

# Audio Utilities
librosa==0.10.1
soundfile==0.12.1